+++ src/rex.asynctask/doc/conf.py
@@ -5,7 +5,7 @@
 
 project = 'rex.asynctask'
 html_title = "REX.ASYNCTASK Documentation"
-extensions = ['sphinx.ext.autodoc']
+extensions = ['sphinx.ext.autodoc', 'sphinxcontrib.autorex']
 master_doc = 'index'
 default_role = 'obj'
 autodoc_default_flags = ['members']